OPSMN-3854: Deserialised the time since UNIX epoch into long type that can further be used for creating date objects.
As part of fixing OPSMN-3808, we implemented a custom date deserialiser
that can handle date strings like "yyyy-MM-dd". For backward compatibility,
we also implemented deserialising time since UNIX epoch into date objects.
The later one had a problem in that the input value was deserialised into
int, which was causing range error. Fixed the deserialiser by using
variable of long type.

+++ WEB-INF/src/com/krishagni/catissueplus/core/common/util/JsonDateDeserializer.java
@@ -33,7 +33,7 @@ public class JsonDateDeserializer extends UntypedObjectDeserializer {
 			}
 		} else if (p.getCurrentToken() == JsonToken.VALUE_NUMBER_INT) {
 			try {
-				ret = new Date(p.getIntValue());
+				ret = new Date(p.getLongValue());
 			} catch (Exception e) {
 				ret = super.deserialize(p, ctxt);
 			}
